Add LazyLoader for ObjectStorageApi
