+++ tests/unit/api/test_utils.py
@@ -14,7 +14,7 @@
 # License along with this library.
 
 import unittest
-from oio.api.object_storage import obj_range_to_meta_chunk_range
+from oio.common.storage_functions import obj_range_to_meta_chunk_range
 
 
 class TestUtils(unittest.TestCase):
